Updated simplified interface
There are many simplified interface changes to mention here. First of
all, the main functinos have docstrings. Secondly, there is a new one
which takes in a vector of times tspan and retruns the solution at every
internal step and at the tspan steps (tspan[0] being the initial
condition, matching ODEInterface, DifferentialEquations.jl, and ODE.jl
APIs). This is the function which is used in the
DifferentialEquations.jl wrapping. An integrator option is given in the
function which allows for switching between BDF and Adams integrators.
Lastly, the default reltol was changed to 1e-3 to be the standard
default (as seen in DifferentialEquations.jl, ODEInterface.jl, MATLAB,
SciPy, etc... anything which is not ODE.jl)