Add IDA to the common interface
This PR implements the necessary algorithm and `solve` function to have IDA match the common interface for DAEs. The DAE solve function is similar to the ODE solve one, but there are some distinct differences in variable ordering in some of the IDA vs CVode functions which makes it too difficult to put them together (and there are some different checks that need to be done, etc.) so it's better to keep them separate. This will require the newest DiffEqBase tag.