Remove double free
The GC is already going to be freeing this so the `empty!` call is unnecessary. The bigger issue is that this can free the memory, and then the GC can try and free the memory, and if those collide it can segfault. `empty!` does have a check in there to try and make that not occur, but that seems to be failing with the newer Sundials binaries, so it's probably better/safer to just use the tie-in to the GC anyways.