Workaround for N_Vector segfaults
Issue https://github.com/SciML/Sundials.jl/issues/367

This fixes a problem seen with Julia >= 1.8, where a temporary NVector was
created and the N_Vector pointer it contains passed to ccall, which is unsafe as
the temporary NVector may be garbage collected.

The fix is to remove convert(N_Vector, x) and replace with the combination of
cconvert and unsafe_convert. An NVector can new be supplied as an
argument to ccall with memory management then handled correctly (
and it is now impossible to explicitly create an N_Vector).

Changes:
 - change to cconvert / unsafe_convert is in src/nvector_wrapper.jl .
 - A global edit (~400 changes) to the autogenerated code in
lib/libsundials_api.jl.
 - small number of changes elsewhere, replacing N_Vector with NVector

NB: The correct fix would be to update the code generator in gen/generate.jl,
and regenerate the wrappers.

NB: As the comments in src/nvector_wrapper.jl show this was an intentional
and (now) incorrect use of temporary variables to avoid garbage collection,
it is possible there are other similar problems (pointers from other types of
temporary variables passed to ccall).