Perform bi-directional sync of integrator u/du values in DAE init
Sundials.jl got an implementation of the OrdinaryDiffEq DAE initialization
interface in #401. Unfortunately, it turns out that this doesn't quite work,
because the u/du in Sundials' integrator struct are not the actual memory
that IDA uses (it has an internal copy). In [1], I propose re-using `u_modified`
to inform us that the shadow copy was modified during initialization. This
is the Sundials side of that interface.

[1] https://github.com/SciML/OrdinaryDiffEq.jl/pull/1969