Correct initial saved value for `dures`
We were saving `u` twice rather than `du` at the start. While we're
at it, also defer the choice of the interpolator to the SciMLBase
code (which does the same thing currently, but let's have this in
one place and give the SciMLBase constructor access to the `du`
vector if we have it).