Mark `SubArray`s as invalid outputs for `IDAIntegrator(out, t)`
Let's imagine we have a matrix that we are filling out, column by column with
`integrator(t)`, but we realize we want to use the non-allocating version.
We try to use `integrator(out[:, idx], t)`, and then realize that of course we
need to make that a view because otherwise slicing just creates a copy.
Imagine our surprise when the view doesn't work either.  Indeed:

```julia
julia> out = zeros(3)
       outv = @view(out[:])
       nv = convert(Sundials.NVector, outv)
       nv[1] = 1.0
       out
3-element Vector{Float64}:
 0.0
 0.0
 0.0
```

The issue, as it turns out, is due to the `convert(NVector, x)` overload
automatically making a copy of non-plain `Vector` types.  I understand
the reasoning behind this, as the backing memory store needs to be
contiguous memory.  It may be possible to write a correct dispatch for
`SubArray`s that are contiguous vectors, but I'm not sure how to do
that, and so have opted instead to just make this an error rather than
silently return untouched memory.